Cleaned up warnings from Xcode code analysis.

+++ source/RKMoment.c
@@ -363,7 +363,7 @@ static void *momentCore(void *in) {
     //        [ t0 - t1 ]
     // [    t0 - t2     ]
     //
-    uint32_t tic = me->tic;
+    uint64_t tic = me->tic;
 
     RKModuloPath path;
     RKPulse *S, *E, *pulses[RKMaxPulsesPerRay];
